client/src/app/components/home/home.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { HomeComponent } from './home.component';
import { RouterTestingModule } from '@angular/router/testing';
import { HttpClientModule } from '@angular/common/http';

describe('HomeComponent', () => {
  let component: HomeComponent;
  let fixture: ComponentFixture<HomeComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ HomeComponent ],
      providers:[],
      imports:[
        RouterTestingModule,
        HttpClientModule,
      ]
    })
    .compileComponents();
    let store = {};
    const mockLocalStorage = {
      getItem: (key: string): string => {
        return key in store ? store[key] : null;
      },
      setItem: (key: string, value: string) => {
        store[key] = `${value}`;
      },
      removeItem: (key: string) => {
        delete store[key];
      },
      clear: () => {
        store = {};
      }
    };
    spyOn(localStorage, 'getItem').and.callFake(mockLocalStorage.getItem);
    spyOn(localStorage, 'setItem').and.callFake(mockLocalStorage.setItem);
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(HomeComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should check a user is log in', () => {
    let user = {
      email: "Test@gmail.com",
      name: "Test",
      admin: true,
      uid: "123456789ASDDFG",
    };
    localStorage.setItem('user', JSON.stringify(user));
    component.identity;
    expect(component.identity).toEqual(true);
  });

  it('should check a user is not log in', () => {
    localStorage.setItem('user',null);
    component.identity;
    expect(component.identity).toEqual(false);
  });
  
});
